requirements.txt
pandas
scikit-learn
numpy
torch
torchvision
fastapi
streamlit
matplotlib
pillow
